requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
crispy-tailwind==0.5.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.7
django-bootstrap4==3.0.1
django-crispy-forms==1.13.0
django-filter==21.1
django-flatpickr==1.0.1
django-import-export==2.6.1
django-widget-tweaks==1.4.8
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
pytz==2021.3
PyYAML==5.4.1
soupsieve==2.2.1
sqlparse==0.4.2
tablib==3.0.0
xlrd==2.0.1
xlwt==1.3.0
